chatbot/trace/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus
prometheus_client
litellm
celery
redis
gevent
evidently[llm]
numpy==2.0.2
langfuse
fastapi
uvicorn